restructure files for better organization
