+++ src/contracts/common/search.query.ts
@@ -1,4 +1,3 @@
-// search.query.ts
 import { IsString, IsOptional } from "class-validator";
 import { ApiPropertyOptional } from "@nestjs/swagger";
