+++ src/contracts/user/user.body.ts
@@ -14,7 +14,6 @@ export class UserBody {
 	public lastName: string;
 
 	@Expose()
-	// We can start adding validation decorators that specify exactly what we expect from the object we will be validating
 	@IsEmail()
 	public email: string;
