+++ src/controllers/users/handlers/get.handler.ts
@@ -1,6 +1,6 @@
 import { NotFoundException } from "@nestjs/common";
 import { prisma } from "../../../lib/prisma";
-import { UserView } from "../../../contracts/user.view";
+import { UserView } from "../../../contracts/user/user.view";
 import { plainToInstance } from "class-transformer";
 
 export const get = async (id: string): Promise<UserView> => {
