+++ src/controllers/users/handlers/getList.handler.ts
@@ -1,6 +1,6 @@
 import { plainToInstance } from "class-transformer";
 import { prisma } from "../../../lib/prisma";
-import { UserView } from "../../../contracts/user.view";
+import { UserView } from "../../../contracts/user/user.view";
 
 export const getList = async (search?: string): Promise<UserView[]> => {
     const where = search
